support device mode
