+++ packages/electron-client-ipc/src/events/gatewayConnection.ts
@@ -1,4 +1,9 @@
-export type GatewayConnectionStatus = 'connected' | 'connecting' | 'disconnected' | 'reconnecting';
+export type GatewayConnectionStatus =
+  | 'connected'
+  | 'connecting'
+  | 'disconnected'
+  | 'reconnecting'
+  | 'authenticating';
 
 export interface GatewayConnectionBroadcastEvents {
   gatewayConnectionStatusChanged: (params: { status: GatewayConnectionStatus }) => void;
