+++ src/server/services/agentRuntime/AgentRuntimeService.ts
@@ -1468,9 +1468,7 @@ export class AgentRuntimeService {
 
     // Create streaming executor context
     const executorContext: RuntimeExecutorContext = {
-      activeDeviceId: metadata?.activeDeviceId,
       agentConfig: metadata?.agentConfig,
-      deviceSystemInfo: metadata?.deviceSystemInfo,
       discordContext: metadata?.discordContext,
       userTimezone: metadata?.userTimezone,
       evalContext: metadata?.evalContext,
